+++ apps/frontend/src/routes/deposits/listing.tsx
@@ -215,7 +215,7 @@ export const columns: ColumnDef<DepositTableType>[] = [
       )
       const sum =
         articles?.reduce(
-          (acc, article) => acc + parseInt(`${article.price}`),
+          (acc, article) => acc + parseFloat(`${article.price}`),
           0,
         ) ?? 0
 
@@ -286,8 +286,10 @@ function DepositDataTableHeaderAction({
 function DepositsSummary() {
   const articles = useLiveQuery(() => db.articles.toArray())
   const total =
-    articles?.reduce((acc, article) => acc + parseInt(`${article.price}`), 0) ??
-    0
+    articles?.reduce(
+      (acc, article) => acc + parseFloat(`${article.price}`),
+      0,
+    ) ?? 0
   const count = articles?.length ?? 0
 
   return (
